requirements.txt



djangorestframework-simplejwt==4.7.2